JDBC connectivity in Eclipse IDE
This is an example for how to create a web application with JDBC connectivity in Eclipse IDE using JavaEE. Here when a user goes to the login.html  page and enter his username and password, it checks whether the user details in the database. If the details are already in the database it redirects to the home.html page. If not, it redirects to the registration.html page and let the user to insert the details about him/her. After the registration process, it goes to the insert_message.html page and says "Successfully inserted".