Update to identify the profession of the user
Now when user login to the website, In the home page it's "welcome to ABC company home page" and "Hi <profession type> and your are a <profession> from <location>" . Also now there is a footer in home page and pages after that. When user click the library link, he is redirected to the library page and there he will see the books that are only matching for his profession.